requirements.txt
Flask==1.1.2
flask-login==0.4.1
gunicorn==20.0.4
psycopg2-binary
flask-sslify==0.1.5
dbfread==2.0.7
bcrypt==3.2.0
spotipy==2.17.1
stravalib==0.10.2
datetime
pytz
requests
urllib3==1.26.4
numpy==1.21.2

